requirements.txt
PyJWT
passlib
bcrypt
boto3
spacy
fastapi
python-dotenv
psycopg2-binary
pydantic
pandas
openpyxl
reportlab
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
blis==1.3.3
boto3==1.42.94
botocore==1.42.94
catalogue==2.0.10
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
cloudpathlib==0.23.0
colorama==0.4.6
confection==1.3.3
cymem==2.0.13
et_xmlfile==2.0.0
fastapi==0.136.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
Jinja2==3.1.6
jmespath==1.1.0
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
numpy==2.4.4
openpyxl==3.1.5
packaging==26.1
pandas==3.0.2
pillow==12.2.0
preshed==3.0.13
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
reportlab==4.4.10
requests==2.33.1
rich==15.0.0
s3transfer==0.16.1
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart_open==7.6.0
spacy==3.8.13
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.49
srsly==2.5.3
starlette==1.0.0
thinc==8.3.13
tqdm==4.67.3
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.45.0
wasabi==1.1.3
weasel==1.0.0
wrapt==2.1.2
